sampling frequency in all EMA file types

+++ src/main/java/org/m2ci/msp/ema/AG501PosFile.java
@@ -15,7 +15,7 @@ public class AG501PosFile extends AG500PosFile {
 	}
 
 	@Override
-	public int getSamplingFrequency() {
+	public double getSamplingFrequency() {
 		return header.getSamplingFrequency();
 	}
